Adding copyrights and moving remaining 2 object classes as engine objects
