+++ Node2Code_Engine/Node2Code_Engine.csproj
@@ -111,6 +111,8 @@
     <Compile Include="Objects\BodyState.cs" />
     <Compile Include="Objects\NodeState.cs" />
     <Compile Include="Objects\ReceiverState.cs" />
+    <Compile Include="Objects\StatementInfo.cs" />
+    <Compile Include="Objects\Variable.cs" />
     <Compile Include="Query\OutputVariable.cs" />
     <Compile Include="Query\Children.cs" />
     <Compile Include="Query\Expression.cs" />
@@ -126,13 +128,6 @@
     <Compile Include="Query\ArgumentValue.cs" />
     <Compile Include="Query\ReturnType.cs" />
     <Compile Include="Query\MethodParameters.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Node2Code_oM\Node2Code_oM.csproj">
-      <Project>{e35fd3cf-5b58-49c2-b44a-e97f656f009b}</Project>
-      <Name>Node2Code_oM</Name>
-      <Private>False</Private>
-    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
